vendor_classification/requirements.txt
# Web Framework
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
pydantic-settings==2.0.3
python-multipart==0.0.6
email-validator==2.0.0

# Database
sqlalchemy==2.0.20
alembic==1.12.0
psycopg2-binary==2.9.7

# Task Queue
celery==5.3.4
redis==4.6.0

# File Processing
numpy==1.24.4  # Pinning NumPy version that's compatible with pandas 2.1.0
pandas==2.1.0
openpyxl==3.1.2
xlsxwriter==3.1.2

# API Clients
openai==0.28.0
httpx==0.24.1

# Security
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1

# Testing
pytest==7.4.2
pytest-asyncio==0.21.1

# Utilities
python-dotenv==1.0.0
tenacity==8.2.3
loguru==0.7.0